requirements.txt
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.5
faiss-cpu==1.8.0
unstructured==0.14.0
pypdf==4.2.0
tiktoken==0.7.0